# infrastructure/rest-api/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Service API
  version: 1.0.0

paths:
  /api/v1/items:
    post:
      summary: Create item
      operationId: createItem
      tags:
        - items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/CreateItemRequest.yaml"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "./schemas/ItemResponse.yaml"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      summary: List items
      operationId: listItems
      tags:
        - items
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas/ItemResponse.yaml"

  /api/v1/items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/ItemId"
    get:
      summary: Get item
      operationId: getItem
      tags:
        - items
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "./schemas/ItemResponse.yaml"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete item
      operationId: deleteItem
      tags:
        - items
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  /api/v1/items/{itemId}/enabled:
    parameters:
      - $ref: "#/components/parameters/ItemId"
    put:
      summary: Set item enabled state
      operationId: setItemEnabled
      tags:
        - items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/SetEnabledRequest.yaml"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "./schemas/ItemResponse.yaml"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  parameters:
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "./schemas/ErrorResponse.yaml"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "./schemas/ErrorResponse.yaml"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "./schemas/ErrorResponse.yaml"